Lucas Commit 18-02-16
Road, Map and intersection are done and javadocced. There might be some issues here and there when we eventually get to testing, that's because I wasnt sure how to handle some things but we will figure these out once our project gets more concrete. Also added a TrafficLights class because It was cleaner.